Q   |   A
------|-------
Bug fix? |	no
New feature? |	no
BC breaks? |	no
Deprecations? |	no
Tests pass? | yes
Fixed tickets | no
License | no
Doc PR | no

Test Purposes: http://www.advanced-javascript.com/2012/11/php-switch-vs-inarray-vs-isset.html

For the Switch took: 2.094s. (3.726s)
For the In_Array took: 1.617s. (2.414s)
For the IsSet took: 0.848s. (0.899s.)