@joshlopes 2015-05-29
yeah -- its just that i am comparing this project with mongo time to time and since there is that small difference i though that would be wiser make it more similar in a proper way :)

Thanks!